Refactors download table statuses into components
