--- components/form-builder/app/DownloadTable.tsx
@@ -1,268 +1,0 @@
-import React from "react";
-import { VaultSubmissionList } from "@lib/types";
-import { useTranslation } from "react-i18next";
-import { ExclamationText } from "./shared";
-import { SkipLinkReusable } from "@components/globals/SkipLinkReusable";
-
-export enum VaultStatus {
-  NEW = "New",
-  DOWNLOADED = "Downloaded",
-  CONFIRMED = "Confirmed",
-  PROBLEM = "Problem",
-}
-
-interface DownloadTableProps {
-  submissions: VaultSubmissionList[];
-  checkedItems: Map<string, boolean>;
-  setCheckedItems: React.Dispatch<React.SetStateAction<Map<string, boolean>>>;
-  getCheckedItems: () => Map<string, boolean>;
-  handleDownload: React.MouseEventHandler<HTMLButtonElement>;
-}
-
-export const DownloadTable = ({
-  submissions,
-  checkedItems,
-  setCheckedItems,
-  getCheckedItems,
-  handleDownload,
-}: DownloadTableProps) => {
-  const { t } = useTranslation("form-builder");
-  const DOWNLOAD_OVERDUE = 15;
-  const CONFIRM_OVERDUE = 15;
-
-  function getDaysPassed(date: Date): number {
-    const dateCreated = new Date(date);
-    const dateToday = new Date();
-    const dateDiff = Math.abs(Number(dateToday) - Number(dateCreated));
-    const daysPassed = Math.ceil(dateDiff / (1000 * 60 * 60 * 24));
-    return daysPassed;
-  }
-
-  // Format date for: DD/MM/YYYY
-  function formatDate(date: Date): string {
-    const dateObj = new Date(date);
-    const day = String(dateObj.getDate()).length <= 2 ? `0${dateObj.getDate()}` : dateObj.getDate();
-    const month =
-      String(dateObj.getMonth()).length <= 2
-        ? `0${dateObj.getMonth() + 1}`
-        : dateObj.getMonth() + 1;
-    const year = dateObj.getFullYear();
-    return `${day}/${month}/${year}`;
-  }
-
-  function formatStatus(vaultStatus: string) {
-    switch (vaultStatus) {
-      case VaultStatus.NEW:
-        return (
-          <span className="p-2 bg-[#ecf3eb] text-[#0c6722]">
-            {t("downloadResponsesTable.status.new")}
-          </span>
-        );
-      case VaultStatus.DOWNLOADED:
-        return (
-          <span className="p-2 bg-[#dcd6fe]">{t("downloadResponsesTable.status.downloaded")}</span>
-        );
-      case VaultStatus.CONFIRMED:
-        return (
-          <span className="p-2 bg-[#e2e8ef]">{t("downloadResponsesTable.status.confirmed")}</span>
-        );
-      case VaultStatus.PROBLEM:
-        return (
-          <span className="p-2 bg-[#f3e9e8] text-[#bc3332]">
-            {t("downloadResponsesTable.status.problem")}
-          </span>
-        );
-      default:
-        return (
-          <span className="p-2 bg-[#f3e9e8] text-[#bc3332]">
-            {t("downloadResponsesTable.unknown")}
-          </span>
-        );
-    }
-  }
-
-  function formatDownloadResponse({
-    vaultStatus,
-    createdAt,
-    downloadedAt,
-  }: {
-    vaultStatus: string;
-    createdAt?: Date;
-    downloadedAt?: Date;
-  }) {
-    if (vaultStatus === "New" && createdAt) {
-      const daysPassed = getDaysPassed(createdAt);
-      const daysLeft = DOWNLOAD_OVERDUE - daysPassed;
-      if (daysLeft > 0) {
-        return t("downloadResponsesTable.status.withinXDays", { daysLeft });
-      }
-      return <ExclamationText text={t("downloadResponsesTable.status.overdue")} />;
-    }
-
-    if (
-      (vaultStatus === VaultStatus.DOWNLOADED ||
-        vaultStatus === VaultStatus.CONFIRMED ||
-        vaultStatus === VaultStatus.PROBLEM) &&
-      downloadedAt
-    ) {
-      return formatDate(downloadedAt);
-    }
-
-    return t("downloadResponsesTable.unknown");
-  }
-
-  function formatConfirmReceipt({
-    vaultStatus,
-    createdAtDate,
-  }: {
-    vaultStatus: string;
-    createdAtDate: Date;
-  }) {
-    switch (vaultStatus) {
-      case VaultStatus.NEW:
-        return t("downloadResponsesTable.status.unconfirmed");
-      case VaultStatus.CONFIRMED:
-        return t("downloadResponsesTable.status.one");
-      case VaultStatus.PROBLEM:
-        return (
-          <span className="p-2 bg-[#f3e9e8] text-[#bc3332] font-bold">
-            {t("downloadResponsesTable.Status.PROBLEM")}
-          </span>
-        );
-      case VaultStatus.DOWNLOADED: {
-        const daysPassed = getDaysPassed(createdAtDate);
-        const daysLeft = CONFIRM_OVERDUE - daysPassed;
-        if (daysLeft > 0) {
-          return t("downloadResponsesTable.status.withinXDays", { daysLeft });
-        }
-        return <ExclamationText text={t("downloadResponsesTable.status.overdue")} />;
-      }
-      default:
-        return t("downloadResponsesTable.unknown");
-    }
-  }
-
-  function formatRemoval({ vaultStatus, removedAt }: { vaultStatus: string; removedAt?: Date }) {
-    if (vaultStatus === VaultStatus.CONFIRMED && removedAt) {
-      const daysLeft = getDaysPassed(removedAt);
-      return t("downloadResponsesTable.status.withinXDays", { daysLeft });
-    }
-
-    if (vaultStatus === VaultStatus.PROBLEM) {
-      return t("downloadResponsesTable.status.wontRemove");
-    }
-
-    if (vaultStatus === VaultStatus.NEW || vaultStatus === VaultStatus.DOWNLOADED) {
-      return t("downloadResponsesTable.status.notSet");
-    }
-
-    return t("downloadResponsesTable.unknown");
-  }
-
-  const handleChecked = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const name = e.target.id;
-    const checked: boolean = e.target.checked;
-    // Note: "new Map" Needed to clone and set so React change detection notices a change in the Map
-    setCheckedItems(new Map(checkedItems.set(name, checked)));
-  };
-
-  return (
-    <section>
-      <SkipLinkReusable
-        text={t("downloadResponsesTable.skipLink")}
-        anchor="#downloadTableButtonId"
-      />
-      <table className="text-sm">
-        <caption className="sr-only">{t("downloadResponsesTable.header.tableTitle")}</caption>
-        <thead className="border-b-2 border-[#6a6d7b]">
-          <tr>
-            <th className="p-4  text-center">{t("downloadResponsesTable.header.select")}</th>
-            <th className="p-4 text-left">{t("downloadResponsesTable.header.number")}</th>
-            <th className="p-4 text-left">{t("downloadResponsesTable.header.status")}</th>
-            <th className="p-4 text-left whitespace-nowrap">
-              {t("downloadResponsesTable.header.downloadResponse")}
-            </th>
-            <th className="p-4 text-left whitespace-nowrap">
-              {t("downloadResponsesTable.header.lastDownloadedBy")}
-            </th>
-            <th className="p-4 text-left whitespace-nowrap">
-              {t("downloadResponsesTable.header.confirmReceipt")}
-            </th>
-            <th className="p-4 text-left">{t("downloadResponsesTable.header.removal")}</th>
-          </tr>
-        </thead>
-        <tbody>
-          {submissions.map((submission, index) => (
-            <tr
-              key={index}
-              className={
-                "border-b-2 border-grey" +
-                (checkedItems.get(submission.name) ? " bg-[#fffbf3]" : "")
-              }
-            >
-              <td className="p-4 flex">
-                {/* TODO: Replace below with Design System checkbox */}
-                <div className="form-builder">
-                  <div className="multiple-choice-wrapper">
-                    <input
-                      id={submission.name}
-                      className="multiple-choice-wrapper"
-                      name="responses"
-                      type="checkbox"
-                      checked={checkedItems.get(submission.name)}
-                      onChange={handleChecked}
-                    />
-                    <label htmlFor={submission.name}>
-                      <span className="sr-only">{submission.name}</span>
-                    </label>
-                  </div>
-                </div>
-              </td>
-              <td className="p-4 whitespace-nowrap">{submission.name}</td>
-              <td className="p-4">{formatStatus(submission.status)}</td>
-              <td className="p-4">
-                {formatDownloadResponse({
-                  vaultStatus: submission.status,
-                  createdAt: submission.createdAt,
-                  downloadedAt: submission.downloadedAt,
-                })}
-              </td>
-              <td className="p-4">
-                <div className="truncate w-48">
-                  {submission.lastDownloadedBy || t("downloadResponsesTable.status.notDownloaded")}
-                </div>
-              </td>
-              <td className="p-4">
-                {formatConfirmReceipt({
-                  vaultStatus: submission.status,
-                  createdAtDate: submission.createdAt,
-                })}
-              </td>
-              <td className="p-4">
-                {formatRemoval({
-                  vaultStatus: submission.status,
-                  removedAt: submission.removedAt,
-                })}
-              </td>
-            </tr>
-          ))}
-        </tbody>
-      </table>
-      <div className="mt-8">
-        {/* NOTE: check/unchek item announcement may be enough for users and additionally announcing
-            the updated Button items checked count may be too verbose. Remove live-region if so */}
-        <button
-          id="downloadTableButtonId"
-          className="gc-button whitespace-nowrap w-auto"
-          type="button"
-          onClick={handleDownload}
-          aria-live="polite"
-        >
-          {t("downloadResponsesTable.download.downloadXSelectedResponses", {
-            size: getCheckedItems().size,
-          })}
-        </button>
-      </div>
-    </section>
-  );
-};
